Added slack samples
in this case of using this samples, you must prepare the Slack App.
on the API site, you can create Slack App and get the Webhook-URL or Access-Token.
* Slack API site: api.slack.com/

and, if you want to use this samples with a 'Grove - I2C High Accuracy Temp&Humi Sensor (SHT35)',
you must also install 'GroveDriverPack'.